# Speed up URI parsing