… src/precompile.jl
… test/uri.jl