Comparison against URIParser
```julia
julia> @benchmark HTTP.URI("http://google.com/")
BenchmarkTools.Trial:
  memory estimate:  336.00 bytes
  allocs estimate:  8
  --------------
  minimum time:     290.481 ns (0.00% GC)
  median time:      299.969 ns (0.00% GC)
  mean time:        368.044 ns (14.65% GC)
  maximum time:     17.551 μs (97.28% GC)
  --------------
  samples:          10000
  evals/sample:     262
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark URIParser.URI("http://google.com/")
BenchmarkTools.Trial:
  memory estimate:  832.00 bytes
  allocs estimate:  14
  --------------
  minimum time:     1.047 μs (0.00% GC)
  median time:      1.095 μs (0.00% GC)
  mean time:        1.236 μs (8.42% GC)
  maximum time:     365.459 μs (98.54% GC)
  --------------
  samples:          10000
  evals/sample:     10
  time tolerance:   5.00%
  memory tolerance: 1.00%

```